UHF-13312: Move chunk merging to chunker class
